moving components & updating typescript
