+++ package.json
@@ -69,9 +69,6 @@
     "@angular/platform-browser-dynamic": "2.2.1",
     "@angular/platform-server": "2.2.1",
     "@angular/router": "3.2.1",
-    "@angularclass/conventions-loader": "1.0.2",
-    "@angularclass/hmr": "~1.2.2",
-    "@angularclass/hmr-loader": "~3.0.2",
     "@ngrx/core": "1.2.0",
     "@ngrx/effects": "2.0.0",
     "@ngrx/store": "2.2.1",
@@ -86,16 +83,11 @@
     "zone.js": "~0.6.26"
   },
   "devDependencies": {
-    "@types/hammerjs": "2.0.33",
-    "@types/jasmine": "2.2.34",
-    "@types/node": "6.0.38",
-    "@types/selenium-webdriver": "2.53.33",
-    "@types/source-map": "0.1.27",
-    "@types/uglify-js": "2.0.27",
-    "@types/webpack": "1.12.34",
+    "@types/jasmine": "2.5.40",
+    "@types/node": "^6.0.54",
     "angular-router-loader": "0.4.0",
     "angular2-template-loader": "0.6.0",
-    "awesome-typescript-loader": "2.2.1",
+    "awesome-typescript-loader": "3.0.0-beta.17",
     "codelyzer": "~1.0.0-beta.4",
     "copy-webpack-plugin": "4.0.0",
     "css-loader": "0.26.0",
@@ -116,7 +108,6 @@
     "karma-sourcemap-loader": "0.3.7",
     "karma-webpack": "1.8.0",
     "parse5": "2.2.2",
-    "protractor": "4.0.10",
     "raw-loader": "0.5.1",
     "rimraf": "~2.5.4",
     "script-ext-html-webpack-plugin": "1.3.2",
@@ -126,10 +117,10 @@
     "to-string-loader": "1.1.4",
     "ts-helpers": "1.1.2",
     "ts-node": "1.7.0",
-    "tslint": "~3.15.1",
+    "tslint": "4.2.0",
     "tslint-loader": "3.3.0",
     "typedoc": "0.5.0",
-    "typescript": "2.0.6",
+    "typescript": "2.1.4",
     "url-loader": "0.5.7",
     "v8-lazy-parse-webpack-plugin": "0.3.0",
     "webpack": "2.1.0-beta.27",
