+++ src/app/app.module.ts
@@ -1,25 +1,19 @@
 import { NgModule, ApplicationRef } from '@angular/core';
+
 import { BrowserModule } from '@angular/platform-browser';
 import { FormsModule } from '@angular/forms';
 import { HttpModule } from '@angular/http';
 import { RouterModule } from '@angular/router';
 import { StoreModule } from '@ngrx/store';
-import { getReducers, addReducer } from "./state";
-
 import { AuthModule } from "./security/index";
+import { EffectsModule } from '@ngrx/effects';
 
-/*
- * Platform and Environment providers/directives/pipes
- */
+import { getReducers } from "./state";
 import { HTTP_PROVIDER_OVERRIDES } from "./data/index";
 import { ROUTES } from './app.routes';
-// App is our top level component
-import { AppComponent } from './app.component';
-import { HomeComponent } from './home';
-import { NoContentComponent } from './no-content';
+import { COMPONENTS, APP_COMPONENT } from "./components";
 import { AppSettings } from './data';
 import { VALIDATORS } from "./forms/index";
-import { EffectsModule } from '@ngrx/effects';
 import { AuthEffects } from "./security/state/auth.effects";
 
 // Application wide providers
@@ -31,11 +25,9 @@ const APP_PROVIDERS = [
  * `AppModule` is the main entry point into Angular2's bootstraping process
  */
 @NgModule({
-  bootstrap: [ AppComponent ],
+  bootstrap: [ APP_COMPONENT ],
   declarations: [
-    AppComponent,
-    HomeComponent,
-    NoContentComponent,
+    ...COMPONENTS,
 		...VALIDATORS
   ],
   imports: [ // import Angular's modules
