+++ src/app/app.routes.ts
@@ -1,7 +1,6 @@
 import { Routes, RouterModule } from '@angular/router';
-import { HomeComponent } from './home';
-import { AboutComponent } from './about';
-import { NoContentComponent } from './no-content';
+import { HomeComponent } from './components/home';
+import { NoContentComponent } from './components/no-content';
 import { AuthGuard } from "./security/index";
 
 export const ROUTES: Routes = [
