+++ src/app/components/app/index.ts
@@ -1,0 +1,1 @@
+export * from "./app.component";
