+++ src/app/components/home/home.component.ts
@@ -2,9 +2,7 @@ import { Component } from '@angular/core';
 
 @Component({
   selector: 'gym-home',
-  styleUrls: [ './home.component.css' ],
   templateUrl: './home.component.html'
 })
 export class HomeComponent {
-
 }
