+++ src/app/components/index.ts
@@ -1,0 +1,11 @@
+import { HomeComponent } from "./home";
+import { NoContentComponent } from "./no-content";
+import { AppComponent } from "./app";
+
+export const APP_COMPONENT = AppComponent;
+
+export const COMPONENTS = [
+	HomeComponent,
+	NoContentComponent,
+	AppComponent
+];
