+++ src/polyfills.browser.ts
@@ -40,8 +40,6 @@ if ('production' === ENV) {
 } else {
   // Development
 
-  Error.stackTraceLimit = Infinity;
-
   require('zone.js/dist/long-stack-trace-zone');
 
 }
