+++ tsconfig.json
@@ -17,13 +17,8 @@
       "es6"
     ],
     "types": [
-      "hammerjs",
       "jasmine",
-      "node",
-      "selenium-webdriver",
-      "source-map",
-      "uglify-js",
-      "webpack"
+			"node"
     ]
   },
   "exclude": [
